+++ tests/sdlConverter.spec.ts
@@ -10,36 +10,73 @@ describe("SDLConverter", () => {
 "export const Mutation = objectType({
   name: \\"Mutation\\",
   definition(t) {
-    t.field(\\"createPost\\", {\\"type\\":\\"Post\\"})
-    t.field(\\"registerClick\\", {\\"type\\":\\"Query\\"})
+    t.field(\\"createPost\\", {
+      type: Post,
+      args: {
+        input: arg({
+          type: CreatePostInput,
+          required: true
+        }),
+      },
+    })
+    t.field(\\"registerClick\\", {
+      type: Query,
+      args: {
+        uuid: uuidArg(),
+      },
+    })
   }
 })
 export const Post = objectType({
   name: \\"Post\\",
   definition(t) {
-    t.implements(\\"Node\\")
+    t.implements(Node)
     t.uuid(\\"uuid\\")
-    t.field(\\"author\\", {\\"type\\":\\"User\\"})
-    t.float(\\"geo\\", {\\"list\\":[true,true]})
-    t.float(\\"messyGeo\\", {\\"list\\":[true,false],\\"nullable\\":true})
+    t.field(\\"author\\", { type: User })
+    t.float(\\"geo\\", { list: [true, true] })
+    t.float(\\"messyGeo\\", {
+      list: [true, false],
+      nullable: true,
+    })
   }
 })
 export const Query = objectType({
   name: \\"Query\\",
   definition(t) {
-    t.field(\\"user\\", {\\"type\\":\\"User\\"})
-    t.list.field(\\"posts\\", {\\"type\\":\\"Post\\"})
+    t.field(\\"user\\", { type: User })
+    t.list.field(\\"posts\\", {
+      type: Post,
+      args: {
+        filters: arg({
+          type: PostFilters,
+          required: true
+        }),
+      },
+    })
   }
 })
 export const User = objectType({
   name: \\"User\\",
   definition(t) {
-    t.implements(\\"Node\\")
-    t.string(\\"name\\")
+    t.implements(Node)
+    t.string(\\"name\\", {
+      description: \\"This is a description of a name\\",
+      args: {
+        prefix: stringArg({ description: \\"And a description of an arg\\" }),
+      },
+    })
     t.string(\\"email\\")
-    t.string(\\"phone\\", {\\"nullable\\":true})
-    t.list.field(\\"posts\\", {\\"type\\":\\"Post\\"})
-    t.field(\\"outEnum\\", {\\"nullable\\":true,\\"type\\":\\"SomeEnum\\"})
+    t.string(\\"phone\\", { nullable: true })
+    t.list.field(\\"posts\\", {
+      type: Post,
+      args: {
+        filters: arg({ type: PostFilters }),
+      },
+    })
+    t.field(\\"outEnum\\", {
+      type: SomeEnum,
+      nullable: true,
+    })
   }
 })"
 `);
@@ -77,6 +114,7 @@ export const SomeEnum = enumType({
       .toMatchInlineSnapshot(`
 "export const Node = interfaceType({
   name: \\"Node\\",
+  description: \\"This is a description of a Node\\",
   definition(t) {
     t.id(\\"id\\")
     t.resolveType(() => null)
@@ -88,44 +126,84 @@ export const SomeEnum = enumType({
 
 test("convertSDL", () => {
   expect(convertSDL(EXAMPLE_SDL)).toMatchInlineSnapshot(`
-"export const Mutation = objectType({
+"import { objectType, arg, uuidArg, stringArg, interfaceType, inputObjectType, enumType, scalarType } from 'nexus';
+
+export const Mutation = objectType({
   name: \\"Mutation\\",
   definition(t) {
-    t.field(\\"createPost\\", {\\"type\\":\\"Post\\"})
-    t.field(\\"registerClick\\", {\\"type\\":\\"Query\\"})
+    t.field(\\"createPost\\", {
+      type: Post,
+      args: {
+        input: arg({
+          type: CreatePostInput,
+          required: true
+        }),
+      },
+    })
+    t.field(\\"registerClick\\", {
+      type: Query,
+      args: {
+        uuid: uuidArg(),
+      },
+    })
   }
 })
 export const Post = objectType({
   name: \\"Post\\",
   definition(t) {
-    t.implements(\\"Node\\")
+    t.implements(Node)
     t.uuid(\\"uuid\\")
-    t.field(\\"author\\", {\\"type\\":\\"User\\"})
-    t.float(\\"geo\\", {\\"list\\":[true,true]})
-    t.float(\\"messyGeo\\", {\\"list\\":[true,false],\\"nullable\\":true})
+    t.field(\\"author\\", { type: User })
+    t.float(\\"geo\\", { list: [true, true] })
+    t.float(\\"messyGeo\\", {
+      list: [true, false],
+      nullable: true,
+    })
   }
 })
 export const Query = objectType({
   name: \\"Query\\",
   definition(t) {
-    t.field(\\"user\\", {\\"type\\":\\"User\\"})
-    t.list.field(\\"posts\\", {\\"type\\":\\"Post\\"})
+    t.field(\\"user\\", { type: User })
+    t.list.field(\\"posts\\", {
+      type: Post,
+      args: {
+        filters: arg({
+          type: PostFilters,
+          required: true
+        }),
+      },
+    })
   }
 })
 export const User = objectType({
   name: \\"User\\",
   definition(t) {
-    t.implements(\\"Node\\")
-    t.string(\\"name\\")
+    t.implements(Node)
+    t.string(\\"name\\", {
+      description: \\"This is a description of a name\\",
+      args: {
+        prefix: stringArg({ description: \\"And a description of an arg\\" }),
+      },
+    })
     t.string(\\"email\\")
-    t.string(\\"phone\\", {\\"nullable\\":true})
-    t.list.field(\\"posts\\", {\\"type\\":\\"Post\\"})
-    t.field(\\"outEnum\\", {\\"nullable\\":true,\\"type\\":\\"SomeEnum\\"})
+    t.string(\\"phone\\", { nullable: true })
+    t.list.field(\\"posts\\", {
+      type: Post,
+      args: {
+        filters: arg({ type: PostFilters }),
+      },
+    })
+    t.field(\\"outEnum\\", {
+      type: SomeEnum,
+      nullable: true,
+    })
   }
 })
 
 export const Node = interfaceType({
   name: \\"Node\\",
+  description: \\"This is a description of a Node\\",
   definition(t) {
     t.id(\\"id\\")
     t.resolveType(() => null)
@@ -135,15 +213,21 @@ export const Node = interfaceType({
 export const CreatePostInput = inputObjectType({
   name: \\"CreatePostInput\\",
   definition(t) {
-    t.string(\\"name\\", {\\"required\\":true})
-    t.id(\\"author\\", {\\"required\\":true})
-    t.float(\\"geo\\", {\\"list\\":[false,true],\\"required\\":true})
+    t.string(\\"name\\", { required: true })
+    t.id(\\"author\\", { required: true })
+    t.float(\\"geo\\", {
+      list: [false, true],
+      required: true,
+    })
   }
 });
 export const PostFilters = inputObjectType({
   name: \\"PostFilters\\",
   definition(t) {
-    t.field(\\"order\\", {\\"required\\":true,\\"type\\":\\"OrderEnum\\"})
+    t.field(\\"order\\", {
+      type: OrderEnum,
+      required: true,
+    })
     t.string(\\"search\\")
   }
 });